att: path entidade entregadores
